requirements.txt
# --- FastAPI backend ---
fastapi==0.115.6
uvicorn[standard]==0.34.0

# --- Config / env ---
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1

# --- RAG / orchestration (LangChain 1.x family) ---
langchain==1.2.6
langchain-core==1.2.7
langsmith==0.6.4
langchain-openai==1.1.7
langchain-chroma==1.1.0

# --- Persistence / vector store ---
chromadb==0.5.23

# --- Graph orchestration (optional, but matches your installed stack) ---
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3

# --- Document ingestion helpers (optional but typical) ---
pypdf==5.1.0

# --- Dev/test ---
pytest==8.3.4